fix:huawei touch not working
