+++ apps/reader/src/hooks/index.ts
@@ -11,3 +11,4 @@ export * from './useMobile'
 export * from './useTextSelection'
 export * from './useTranslation'
 export * from './useTypography'
+export * from './useTouchEvent'
